[rename]: Moved components into util and container folder
