+++ frontend/src/components/util/Icon.js
@@ -1,8 +1,8 @@
-import UK from '../images/icons/countries/gb.png';
-import US from '../images/icons/countries/us.png';
-import KENYA from '../images/icons/countries/ke.png';
-import myntIcon from '../images/icons/mynt_icon.png';
-import defaultIcon from '../images/icons/default_icon.png';
+import UK from '../../../public/images/icons/countries/gb.png';
+import US from '../../../public/images/icons/countries/us.png';
+import KENYA from '../../../public/images/icons/countries/ke.png';
+import myntIcon from '../../../public/images/icons/mynt_icon.png';
+import defaultIcon from '../../../public/images/icons/default_icon.png';
 import { Image } from "@chakra-ui/react";
 import { SlArrowLeft, SlArrowRight } from "react-icons/sl";
