+++ frontend/src/pages/AccountPage.js
@@ -3,14 +3,14 @@ import { Box } from '@chakra-ui/react';
 
 import useAxios from '../hooks/useAxios';
 import useFormatAmount from '../hooks/useFormatAmount';
-import Container from '../components/Container';
-import Icon from '../components/Icon';
+import Container from '../components/container/Container';
+import Icon from '../components/util/Icon';
 import CustomText from '../components/CustomText';
 import CustomButton from '../components/forms/CustomButton';
-import ContainerRowBalanceWrapper from '../components/ContainerRowBalanceWrapper';
-import InfoBlock from '../components/InfoBlock';
-import DateTimeDisplay from '../components/DateTimeDisplay';
-import CustomBox from '../components/CustomBox';
+import ContainerRowBalanceWrapper from '../components/container/ContainerRowBalanceWrapper';
+import InfoBlock from '../components/util/InfoBlock';
+import DateTimeDisplay from '../components/util/DateTimeDisplay';
+import CustomBox from '../components/util/CustomBox';
 
 const accountDetail = {
   'id': '1',
