+++ frontend/src/pages/Remittance/Amount.js
@@ -1,9 +1,9 @@
 import React, {useState} from 'react';
 import {useLocation, useNavigate} from 'react-router-dom';
 import CustomHeading from "../../components/CustomHeading";
-import Container from "../../components/Container";
-import Icon from "../../components/Icon";
-import InfoBlock from "../../components/InfoBlock";
+import Container from "../../components/container/Container";
+import Icon from "../../components/util/Icon";
+import InfoBlock from "../../components/util/InfoBlock";
 import CustomText from "../../components/CustomText";
 import CustomForm from "../../components/forms/CustomForm";
 import {Box} from "@chakra-ui/react";
